+++ euchrecli/util/__init__.py
@@ -1,4 +1,4 @@
 from .output_util import output
 from .card_util import create_deck, deal_hand
-from .player_util import rotate_dealer, rotate_trick_order
+from .player_util import unique_cpu_name, rotate_dealer, rotate_trick_order
 from .rule_util import valid_play, hand_winner, trick_winner